Token actions are now always called exactly once
There could've been more action calls of tokens because in case of
reduction, token is not consumed from the input so when you ask for
next_token() after the reduction, the same token is returned.

I've made it so that Parser remembers last token until it is shifted to
the stack so this also makes it more efficient because each token
requires just exaclty one call of RE2.