Don't record visited symbols when calculating includes and lookback relations
It can leave out certain items in states with the same inspected
symbols. This can result in corrupted grammar like the one added in
tests. This grammar failed with syntax error because it didn't expect
that `a` may be followed by `@end`. It only expected `b` as the next
symbol.